Alterado menu e tipo de variáveis

+++ Compras/src/compras/Item.java
@@ -11,12 +11,9 @@ package compras;
  */
 public class Item {
 
-
-
-
     private int Codigo_de_Barras;
     private String Descricao;
-    private int valor;
+    private double valor;
     
     public int getCodigo_de_Barras() {
         return Codigo_de_Barras;
@@ -34,15 +31,15 @@ public class Item {
         this.Descricao = Descricao;
     }
 
-    public int getValor() {
+    public double getValor() {
         return valor;
     }
 
-    public void setValor(int valor) {
+    public void setValor(double valor) {
         this.valor = valor;
     }
     
-    public Item(int Codigo_de_Barras, String Descricao, int Valor){
+    public Item(int Codigo_de_Barras, String Descricao, double Valor){
         this.Codigo_de_Barras = Codigo_de_Barras;
         this.Descricao = Descricao;
         this.valor = Valor;
